1labaBOOKS3.0/app/src/main/java/com/codepath/android/booksearch/net/BookClient.java
package com.codepath.android.booksearch.net;

import android.util.Log;

import com.google.gson.Gson;

import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers;
import io.reactivex.rxjava3.core.Observable;
import io.reactivex.rxjava3.schedulers.Schedulers;
import io.reactivex.rxjava3.subjects.PublishSubject;
import io.reactivex.rxjava3.subjects.Subject;
import retrofit2.Call;
import retrofit2.Callback;
import retrofit2.Response;
import retrofit2.Retrofit;
import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory;
import retrofit2.converter.gson.GsonConverterFactory;


import java.io.UnsupportedEncodingException;
import java.net.URLEncoder;

public class BookClient {
    private static final String API_BASE_URL = "https://openlibrary.org/";
    private final OpenLibraryApi api;

    // Конструктор, принимающий OpenLibraryApi в качестве аргумента
    public BookClient(OpenLibraryApi api) {
        this.api = api;
    }

    // Используем Subject для кэширования

    public BookClient() {
        Retrofit retrofit = new Retrofit.Builder()
                .baseUrl(API_BASE_URL)
                .addConverterFactory(GsonConverterFactory.create())
                .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
                .build();

        api = retrofit.create(OpenLibraryApi.class);
    }

    public Observable<BookResponse> searchBooks(String query) {
        return api.searchBooks(query)
                .subscribeOn(Schedulers.io())
                .observeOn(AndroidSchedulers.mainThread());
    }
}